fix: gu leisure score to Z-Score
