+++ measure/LEI/results/gu_leisure_score.csv
@@ -1,0 +1,26 @@
+﻿자치구,총_행사_수,행사_다양성_지수,여가도_점수_원본,여가도_조정점수
+종로구,434,2.3282338863038747,1444.4535066558817,91.51754335586529
+중구,208,1.9471544142197257,613.0081181577029,62.90520849793569
+마포구,169,2.123873686116747,527.9346529537303,59.97759532259865
+구로구,96,2.0667103503765323,294.4041936361471,51.941166479580986
+은평구,112,1.5767404287493816,288.59492801993076,51.74125358645631
+송파구,110,1.554744962889005,281.02194591779056,51.480646332177194
+서초구,93,1.9394556762229112,273.36937788873075,51.21730030608877
+도봉구,92,1.66933801179442,245.57909708508666,50.26095990039356
+성동구,76,2.2091081177014855,243.8922169453129,50.20290969037126
+노원구,74,2.196889522043106,236.56982463118985,49.95092592360666
+광진구,72,2.07722341452841,221.56008584604552,49.43439930839946
+관악구,63,2.0074367660884387,189.46851626357164,48.330039662279056
+강동구,56,2.0751936300367184,172.21084328205623,47.73615541558467
+동대문구,68,1.3248142278695447,158.08736749512903,47.250127560762664
+강북구,42,2.004975865091968,126.20898633386264,46.15310431770589
+금천구,32,2.184812305085318,101.91399376273017,45.3170464449552
+양천구,31,2.162380407853939,98.0337926434721,45.183517995250455
+강남구,47,0.9229500174990528,90.37865082245548,44.92008339786397
+용산구,31,1.335508794110628,72.40077261742947,44.3014148979804
+중랑구,30,1.3792922544726753,71.37876763418026,44.26624488054845
+영등포구,17,1.7315354082622822,46.4361019404588,43.40789878476069
+동작구,18,1.5316410677211536,45.569539218980765,43.37807796537156
+서대문구,15,1.956463203826685,44.34694805740027,43.336005222899956
+성북구,13,1.8184831630123652,36.64028111916075,43.070797503757106
+강서구,10,1.6434177197931796,26.434177197931795,42.71957724680613
